requirements.txt
openai
google-generativeai
groq
ollama
python-dotenv
colorama
duckduckgo-search
pygments
rich
Pillow
prompt_toolkit
requests